# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API pour le jeu BombParty",
        "title": "BombParty API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:7774",
    "paths": {
...
        "/user/login": {
            "post": {
                "description": "Authentifie un utilisateur avec son email et mot de passe et retourne un token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Connexion utilisateur",
                "parameters": [
                    {
                        "description": "Identifiants de connexion",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserLoginPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token JWT généré",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur avec le payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "Enregistre un nouvel utilisateur avec un nom d'utilisateur, email et mot de passe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Créer un nouveau compte utilisateur",
                "parameters": [
                    {
                        "description": "Informations d'inscription",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserCreatePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token JWT généré",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur avec le payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.UserCreatePayload": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.UserLoginPayload": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" suivi de votre token JWT",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}